add validation and unit testing

+++ ProductAPI/Controllers/ProductController.cs
@@ -1,7 +1,9 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using FluentValidation;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProductBusiness;
 using ProductBusiness.Dtos;
+using ProductBusiness.Validator;
 using ProductData;
 using ProductEntity;
 using System;
@@ -16,10 +18,12 @@ namespace ProductAPI.Controllers
     public class ProductController : Controller
     {
         private readonly IProductService _productService;
+        private ProductValidator _productValidator;
 
         public ProductController(IProductService productService) 
         {
             _productService = productService;
+            _productValidator = new ProductValidator();
         }
 
         [HttpGet("Products")]
@@ -34,6 +38,11 @@ namespace ProductAPI.Controllers
         [HttpGet("{id}", Name = "GetProduct")]
         public async Task<ActionResult<ProductDto>> GetProduct(string id)
         {
+            if (string.IsNullOrWhiteSpace(id)) 
+            {
+                return BadRequest("Invalid Product Id");
+            }
+
             var productToReturn = await _productService.GetProduct(id);
 
             if (productToReturn == null) 
@@ -47,6 +56,14 @@ namespace ProductAPI.Controllers
         [HttpPost("Add")]        
         public async Task<ActionResult<ProductDto>> Create(ProductDto productDto)
         {
+
+            var result = _productValidator.Validate(productDto);
+
+            if (!result.IsValid) 
+            {
+                return BadRequest(result.Errors);
+            }
+
             var createdProductId = await _productService.AddProduct(productDto);
 
             var productToReturn = await _productService.GetProduct(createdProductId);
@@ -55,21 +72,33 @@ namespace ProductAPI.Controllers
         }
 
         [HttpPost("Update")]
-        public async Task<IActionResult> Update(ProductDto productDto)
+        public async Task<ActionResult<ProductDto>> Update(ProductDto productDto)
         {
+            var result = _productValidator.Validate(productDto);
+
+            if (!result.IsValid)
+            {
+                return BadRequest(result.Errors);
+            }
+
             var updatedProductId = await _productService.UpdateProduct(productDto);
 
             var productToReturn = await _productService.GetProduct(updatedProductId);
 
-            return Ok(productToReturn);
+            return productToReturn;
         }
 
         [HttpPost("Delete/{id}")]
         public async Task<IActionResult> Delete(string id)
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                return BadRequest("Invalid Product Id");
+            }
+
             if (await _productService.DeleteProduct(id)) 
             {
-                return NoContent();
+                return Ok("Product Deleted");
             }            
 
             throw new Exception("Error deleting the product");
